\name{semdiag-package}
\alias{semdiag-package}
\alias{semdiag-package}
\docType{package}
\title{
Structural equation modeling diagnostic
}
\description{
Structural equation modeling diagnostic
}
\details{
\tabular{ll}{
Package: \tab semdiag\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-02-07\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
To use the package, see 
}
\author{
Zhiyong Zhang and Ke-Hai Yuan

Maintainer: Zhiyong Zhang <zhiyongzhang@nd.edu>
}
\references{
To be added
}
