\name{semdiag.combinations}
\alias{semdiag.combinations}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Enumerate the Combinations of the Elements of a Vector}
\description{
Enumerate the Combinations of the Elements of a Vector}
\usage{
semdiag.combinations(n, r)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{n}{ Size of the source vector }
  \item{r}{ Size of the target vectors }  
}

