\name{semdiag.cpp}
\alias{semdiag.cpp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Case profile plot
}
\description{Generate profile plots for slected cases}
\usage{
semdiag.cpp(d, cases)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{d}{Object from \code{semdiag}}
  \item{cases}{Cases to be plotted}
}

\author{
Ke-Hai Yuan and Zhiyong Zhang
}
