\name{semdiag.DP}
\alias{semdiag.DP}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generate a duplication matrix
}
\description{Generate a duplication matrix}
\usage{
semdiag.DP(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A matrix
}
}

\author{
Ke-Hai Yuan and Zhiyong Zhang
}
