\name{semdiag.eqs.lisrel}
\alias{semdiag.eqs.lisrel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Format EQS output into LISREL format}
\description{
Format EQS output into LISREL format
}
\usage{
semdiag.eqs.lisrel(model)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{model}{
An EQS output from the function \code{semdiag.run.eqs}
}
}
\value{
 Return LISREL notation matrices
}