\name{semdiag.eqs}
\alias{semdiag.eqs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Orgnize results from EQS
}
\description{
Orgnize results from EQS
}
\usage{
semdiag.eqs(res)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{res}{
EQS output
}
}
\value{
Return model chi-square, p-value and parameter estimates
}
