\name{semdiag.input.model}
\alias{semdiag.input.model}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Input model in R
}
\description{
Input model in R
}
\usage{
semdiag.input.model(file = "")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{
A file name or blank
}
}
\value{
EQS model
}
