\name{semdiag.lisrel}
\alias{semdiag.lisrel}
\alias{semdiag.start}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generate the lisrel notation matrices based on an sem object.
}
\description{Generate the lisrel notation matrices based on an sem object.}
\usage{
semdiag.lisrel(model)

semdiag.start(ram.path, coeff)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{model}{An \code{sem} model}
  \item{ram.path}{An ram path object}
  \item{coeff}{Coefficients to be used as starting values.}
}

\author{
Ke-Hai Yuan and Zhiyong Zhang
}
