\name{semdiag.musig}
\alias{semdiag.musig}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Robust covariance estimation
}
\description{
Robust covariance estimation
}
\usage{
semdiag.musig(x, varphi, max_it = 1000)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Data
}
  \item{varphi}{
Percentage of data to be down-weighted
}
  \item{max_it}{
The maximum number of iterations
}
}

\value{
Estimated mean and covariance matrix
}
