\name{semdiag.orthog}
\alias{semdiag.orthog}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
semdiag.orthog
}
\description{
semdiag.orthog
}
\usage{
semdiag.orthog(A)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{A}{
A matrix
}
}

\value{
Return
}
