\name{semdiag.parse}
\alias{semdiag.parse}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Parse EQS input file
}
\description{
Parse EQS input file
}
\usage{
semdiag.parse(eqs)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{eqs}{
EQS input file
}
}

\value{
RAM notation for the model in EQS input file
}
