\name{semdiag.vech}
\alias{semdiag.vech}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Stacking lower triange of a matrix to a vector
}
\description{Stacking lower triange of a matrix to a vector}
\usage{
semdiag.vech(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A matrix
}
}
