\name{semdiag.write.eqs}
\alias{semdiag.write.eqs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generate an EQS input file
}
\description{
Generate an EQS input file. This is an internal function.
}
\usage{
semdiag.write.eqs(eqs, par, N, P)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{eqs}{EQS input file}
  \item{par}{Parameter estimates}
  \item{N}{Sample size}
  \item{P}{Number of variables}
}

\value{
A filed called eqsonce.eqs.
}
